Move resolveRowMetadata to the shared table processor helper
